port allegro-04 (minimal) changes to v03 and remove v04
