--- FCCee/ALLEGRO/compact/ALLEGRO_o1_v04/DectEmptyMaster.xml
@@ -1,1 +1,0 @@
-../ALLEGRO_o1_v03/DectEmptyMaster.xml
