--- FCCee/ALLEGRO/compact/ALLEGRO_o1_v04/DriftChamber_o1_v02.xml
@@ -1,1 +1,0 @@
-../../../IDEA/compact/IDEA_o1_v03/DriftChamber_o1_v02.xml
