--- FCCee/ALLEGRO/compact/ALLEGRO_o1_v04/HCalBarrel_TileCal_v02.xml
@@ -1,1 +1,0 @@
-../ALLEGRO_o1_v03/HCalBarrel_TileCal_v02.xml
