--- FCCee/ALLEGRO/compact/ALLEGRO_o1_v04/MuonTagger.xml
@@ -1,1 +1,0 @@
-../ALLEGRO_o1_v03/MuonTagger.xml
