--- FCCee/ALLEGRO/compact/ALLEGRO_o1_v04/MuonTaggerPhiTheta.xml
@@ -1,1 +1,0 @@
-../ALLEGRO_o1_v03/MuonTaggerPhiTheta.xml
