--- FCCee/ALLEGRO/compact/ALLEGRO_o1_v04/SiliconWrapper_o1_v03.xml
@@ -1,1 +1,0 @@
-../../../IDEA/compact/IDEA_o1_v03/SiliconWrapper_o1_v03.xml
